backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.35
alembic==1.13.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
playwright>=1.55.0
greenlet>=3.2.4
pyee>=13.0.0
openai>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6
